+++ src/bitset.cpp
@@ -43,6 +43,11 @@ void bitset_remove(
 }
 
 //[[Rcpp::export]]
+void bitset_clear(const Rcpp::XPtr<individual_index_t> b) {
+    b->clear();
+}
+
+//[[Rcpp::export]]
 size_t bitset_size(const Rcpp::XPtr<individual_index_t> b) {
     return b->size();
 }